Refactor the post controller
